# docs/swagger.yaml
definitions:
...
  job_rate.CalculatePaymentRequest:
    properties:
      companyParts:
        type: number
      deduction:
        type: number
      rateFlat:
        type: number
      ratePercent:
        type: number
      salePrice:
        type: number
      techParts:
        type: number
    type: object
  job_rate.CalculatePaymentResponse:
    properties:
      payment:
        type: number
    type: object
  job_rate.CreateRequest:
    properties:
      companyParts:
        type: number
      deduction:
        type: number
      jobRateStatusId:
        type: integer
      notes:
        type: string
      partsReplaced:
        type: string
      rateFlat:
        type: number
      ratePercent:
        type: number
      salePrice:
        type: number
      techParts:
        type: number
      userId:
        type: integer
    type: object
  job_rate.JobRate:
    properties:
      companyParts:
        type: number
      createdAt:
        type: string
      deduction:
        type: number
      deletedAt:
        type: string
      id:
        type: integer
      jobId:
        type: integer
      jobRateStatusId:
        type: integer
      notes:
        type: string
      paid:
        type: boolean
      partsReplaced:
        type: string
      payment:
        type: number
      rateFlat:
        type: number
      ratePercent:
        type: number
      salePrice:
        type: number
      techParts:
        type: number
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  job_rate.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/job_rate.JobRate'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  job_rate.UpdateRequest:
    properties:
      companyParts:
        type: number
      deduction:
        type: number
      jobRateStatusId:
        type: integer
      notes:
        type: string
      paid:
        type: boolean
      partsReplaced:
        type: string
      rateFlat:
        type: number
      ratePercent:
        type: number
      salePrice:
        type: number
      techParts:
        type: number
      userId:
        type: integer
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@jvair.com
    name: API Support
    url: http://www.jvair.com/support
  description: API para el sistema JVAIR
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: JVAIR API
  version: "1.0"
paths:
...
  /api/v1/calculate-rate-payment:
    post:
      consumes:
      - application/json
      description: Calculate payment amount based on rate formula
      parameters:
      - description: Payment calculation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_rate.CalculatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job_rate.CalculatePaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Calculate rate payment
      tags:
      - JobRates
...
  /api/v1/jobs/{jobId}/rates:
    get:
      consumes:
      - application/json
      description: Get all rates for a specific job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job_rate.ListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List job rates
      tags:
      - JobRates
    post:
      consumes:
      - application/json
      description: Create a new rate/commission for a job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - description: Rate data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_rate.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/job_rate.JobRate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create job rate
      tags:
      - JobRates
  /api/v1/jobs/{jobId}/rates/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a rate by ID
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - description: Rate ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete job rate
      tags:
      - JobRates
    put:
      consumes:
      - application/json
      description: Update a rate by ID
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - description: Rate ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rate data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_rate.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job_rate.JobRate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update job rate
      tags:
      - JobRates
...
securityDefinitions:
  BearerAuth:
    description: Ingrese el token con el prefijo 'Bearer '
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...